requirements.txt
numpy
pandas
torch==2.4.1
streamlit
Pillow
opencv-python<=4.9.0.80
ultralytics>=8.2.34
